tnetstring: use qint64 for base int type, and support conversion from others

+++ tnetstring.h
@@ -34,7 +34,7 @@ enum Type
 };
 
 QByteArray fromByteArray(const QByteArray &in);
-QByteArray fromInt(int in);
+QByteArray fromInt(qint64 in);
 QByteArray fromDouble(double in);
 QByteArray fromBool(bool in);
 QByteArray fromNull();
@@ -44,7 +44,7 @@ QByteArray fromVariant(const QVariant &in);
 
 bool check(const QByteArray &in, int offset, Type *type, int *dataOffset, int *dataSize);
 QByteArray toByteArray(const QByteArray &in, int offset, int dataOffset, int dataSize, bool *ok = 0);
-int toInt(const QByteArray &in, int offset, int dataOffset, int dataSize, bool *ok = 0);
+qint64 toInt(const QByteArray &in, int offset, int dataOffset, int dataSize, bool *ok = 0);
 double toDouble(const QByteArray &in, int offset, int dataOffset, int dataSize, bool *ok = 0);
 bool toBool(const QByteArray &in, int offset, int dataOffset, int dataSize, bool *ok = 0);
 void toNull(const QByteArray &in, int offset, int dataOffset, int dataSize, bool *ok = 0);
